Add O(n) preprocessing RMQ implementation
This implementation requires only O(n) preprocessing time
compared to the O(n log n) dynamic programming only implementation.
However, the theoretical time to answer a query is now O(log n),
which is worse than O(1) of the DP implementation.

In practice this still seems to be the fastest algorithm, as shown
by Fischer and Heun (2006).