Assign unique ID to each merged reads node
When writing to GFA2, the corresponding fragments (reads) need to
be in the correct order, with the right segment ID's of the reads
spelling this merged node. However, segments in GFA2 do not have
an orientation, so when writing a merged node, you can not recover
whether this merged node was the original or the reverse.