Improve GFA2 assembly graph reconstruction
This is an improved method to include any singleton nodes in the
assembly graph. Previously, the reconstruction method added
two nodes for each read: one for its reverse complement
and one for the original. But, when dealing with subgraphs of
the whole assembly graph, sometimes you don't want to include both
of them. It is possible that a subgraph involves the original read
but not the reverse complement. The reconstruction process
therefore should not add the reverse complement node.

The current code now checks if any segment given is involved in
some edge, and if not adds the node.