Delay alignments parsing until needed in `phase`
When the `phasm phase` command is used on a subgraph with only a
single linear path, then the phasing algorithm is not applied,
and therefore the pairwise local alignment information is not
needed.

This commit delays the reading of all pairwise local alignments
until it's needed, because reading this information takes a lot
of time.