Fix retrieval of aligning reads
Local aligments between reads are output as follows:

    read1+ (a read) aligns with read4- (b read)

The reverse complement of a read is **never** the a-read. This
induced a bug in PHASM when we try to store all pairwise
alignments between reads. We store it as follows:

    alignments[a_read][b_read] = la_metadata

Now, say read4- is included in the assembly graph, if we want to
know which other reads align to this read, we would ask our dict:
alignments[read4-]. But because the reverse complement is never
a a-read, this dictionary would be empty!